+++ trideni_JHV_v2.5_Lite.py
@@ -20,7 +20,7 @@ def whole_function():
             for dirs in folders: # pole folders uz je filtrovano od ostatnich souboru...
                 if (dirs != folder_name[0]) and (dirs != folder_name[1]) and (dirs != folder_name[2]):
                     number_of_files = 0
-                    if os.path.exists(path + dirs):
+                    if os.path.isdir(path + dirs):
                         for files in os.listdir(path + dirs):
                             number_of_files +=1
                         if number_of_files == 0:
@@ -34,7 +34,7 @@ def whole_function():
         else:
             for dirs in folders: # pole folders uz je filtrovano od ostatnich souboru...
                 number_of_files = 0
-                if os.path.exists(path + dirs):
+                if os.path.isdir(path + dirs):
                     for files in os.listdir(path + dirs):
                         number_of_files +=1
                     if number_of_files == 0:
@@ -89,7 +89,8 @@ def whole_function():
             for i in range(0,len(folders)):
                 for files in os.listdir(path + folders[i]):
                     if ".bmp" in files:
-                        shutil.move(path + folders[i] + "/" + files , path + '/' + files)
+                        if os.path.exists(path + folders[i] + "/" + files):
+                            shutil.move(path + folders[i] + "/" + files , path + '/' + files)
             
         def Get_func_number(file_for_analyze):
             files_split = file_for_analyze.split("&")
@@ -150,12 +151,14 @@ def whole_function():
 
                 if count == len(self.files_type_arr): # overeni zda je od vsech typu souboru jeden
                     ok_count += 1
-                    shutil.move(path + '/' + self.files_arr[i] , path + folder_name[0] + "/" + self.files_arr[i]) #přesun do OK složky
+                    if os.path.exists(path + '/' + self.files_arr[i]):
+                        shutil.move(path + '/' + self.files_arr[i] , path + folder_name[0] + "/" + self.files_arr[i]) #přesun do OK složky
                     count = 0
                     
                 else:
                     nok_count += 1
-                    shutil.move(path + '/' + self.files_arr[i] , path + folder_name[1] + "/" + self.files_arr[i]) #přesun do temp složky
+                    if os.path.exists(path + '/' + self.files_arr[i]):
+                        shutil.move(path + '/' + self.files_arr[i] , path + folder_name[1] + "/" + self.files_arr[i]) #přesun do temp složky
                     count = 0
             
             if error_length == 1:
@@ -305,10 +308,11 @@ def whole_function():
             #presun souboru do slozek:
             if sort_by == 0:
                 for files in os.listdir(path + folder_name[0]): #v OK slozce
-                    for items in folder_name: #pro vsechny slozky...
-                        if items in files:
-                            if not os.path.exists(path + items + "/" + files):
-                                shutil.move(path + folder_name[0] + "/" + files, path + items + "/" + files)
+                    if ".bmp" in files:
+                        for items in folder_name: #pro vsechny slozky...
+                            if items in files:
+                                if not os.path.exists(path + items + "/" + files):
+                                    shutil.move(path + folder_name[0] + "/" + files, path + items + "/" + files)
 
             if sort_by == 4:
                 for files in os.listdir(path + self.pair_folder):
@@ -366,7 +370,7 @@ def whole_function():
         print("Analýza složek... ")
         def sync_folders():
             folders = []
-            unsupported_formats = [".exe",".pdf",".ifz",".bmp",".txt",".v",".xml",".changed",".doc",".docx",".xls",".xlsx",".ppt",".pptx",".csv",".py",".msi",".jpg"]
+            unsupported_formats = [".exe",".pdf",".ifz",".bmp",".txt",".v",".xml",".changed",".doc",".docx",".xls",".xlsx",".ppt",".pptx",".csv",".py",".msi",".jpg",".img",".png"]
             if os.path.exists(path):
                 for files in os.listdir(path):
                     #ignorace ostatnich typu souboru:
